update submodule fmtlib

+++ dbms/src/Storages/Page/V3/WAL/serialize.cpp
@@ -20,6 +20,8 @@
 #include <Storages/Page/V3/PageEntry.h>
 #include <Storages/Page/V3/WAL/serialize.h>
 #include <Storages/Page/WriteBatch.h>
+#include <magic_enum.hpp>
+#include "common/types.h"
 
 namespace DB::PS::V3::ser
 {
@@ -218,7 +220,7 @@ void deserializeFrom(ReadBuffer & buf, PageEntriesEdit & edit)
             break;
         }
         default:
-            throw Exception(fmt::format("Unknown record type: {}", record_type), ErrorCodes::LOGICAL_ERROR);
+            throw Exception(fmt::format("Unknown record type: {}", static_cast<Int32>(record_type)), ErrorCodes::LOGICAL_ERROR);
         }
     }
 }
